Changed application Manafest
Add the following to your application's manifest:

The android.permission.INTERNET permission so the Android application
can send the registration token to the app server.
The android.permission.WAKE_LOCK permission so the application can keep
the processor from sleeping when a message is received.
The com.google.android.c2dm.permission.RECEIVE permission so the Android
application can register and receive messages.
An applicationPackage + ".permission.C2D_MESSAGE" permission to prevent
other Android applications from registering and receiving the Android
application's messages. The permission name must exactly match this
pattern—otherwise the Android application will not receive the messages.
A declaration of GcmReceiver, which handles messages sent from GCM to
your application. Because this service needs permission to receive
messages from GCM, add com.google.android.c2dm.permission.SEND to the
receiver.
A declaration of GcmListenerService, which enables various aspects of
handling messages such as detecting different downstream message types,
determining upstream send status, and automatically displaying simple
notifications on the app’s behalf.
A service that extends InstanceIDListenerService, to handle the
creation, rotation, and updating of registration tokens.
If the GCM feature is critical to the Android application's function, be
sure to set android:minSdkVersion="8" or higher in the manifest. This
ensures that the Android application cannot be installed in an
environment in which it could not run properly.